extract equipment section
